modules: app: Merge with trigger module to simplify application
The app and trigger modules are merged to simplify the
application logic. It's now all in the app module.

This also allows for the environmental module to have only one
state and hence no need for state machine, as triggers will
not be sent if time is not available.
