+++ app/prj.conf
@@ -221,7 +221,6 @@ CONFIG_APP_FOTA_LOG_LEVEL_DBG=y
 CONFIG_APP_BATTERY_LOG_LEVEL_DBG=y
 CONFIG_APP_LOCATION_LOG_LEVEL_DBG=y
 CONFIG_APP_TRANSPORT_LOG_LEVEL_DBG=y
-CONFIG_APP_TRIGGER_LOG_LEVEL_DBG=y
 CONFIG_APP_NETWORK_LOG_LEVEL_DBG=y
 CONFIG_APP_LED_LOG_LEVEL_DBG=y
 CONFIG_APP_ENVIRONMENTAL_LOG_LEVEL_DBG=y
