--- app/src/modules/trigger/CMakeLists.txt
@@ -1,7 +1,0 @@
-#
-# Copyright (c) 2024 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/trigger.c)
