--- app/src/modules/trigger/Kconfig.trigger
@@ -1,17 +1,0 @@
-#
-# Copyright (c) 2024 Nordic Semiconductor
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-menu "Trigger"
-
-config APP_TRIGGER_TIMEOUT_SECONDS
-	int "Trigger timer timeout"
-	default 600
-
-module = APP_TRIGGER
-module-str = Trigger
-source "subsys/logging/Kconfig.template.log_config"
-
-endmenu # Trigger
